Add files via upload
These are the latest files I have to date from Siggy's project. This also includes the work from Week 6 and some newer changes I made to the sprites.

Update:  Commodore 64 Machine Language Project Week 6

Bug issue; Still need to get the left sprite shapes to show. The problem lies in the label called AnimTest2. Darren and I managed to get the sprite image data to show when the player moves the controller to the left, but the right still needs work.

Note: This does not include the scrolling code from Week 5 as Siggy has not uploaded anything to Google Drive yet.